Move and refactor input classes
